pyproject.toml
[tool.poetry]
name = "UStore"
version = "0.1.0"
description = "Ecommerce desarrollado para practica con django"
authors = ["Maynor Pineda <sbethuell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
djoser = "^2.1.0"
django-simple-history = "^3.1.1"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.12.0"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
rich = "^12.4.1"
django-debug-toolbar = "^3.4.0"
ipython = "^8.3.0"
pywatchman = "^1.4.1"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"

# [tool.mypy]
# python_version = '3.10'
# check_untyped_defs = true
# ignore_errors = false
# ignore_missing_imports = true
# strict_optional = true
# warn_unused_ignores = true
# warn_redundant_casts = true
# warn_unused_configs = true
# disallow_incomplete_defs = true
# plugins = ['mypy_django_plugin.main']

# [[tool.mypy.overrides]]
# module = '*.migrations.*'
# ignore_errors = true

# [[tool.mypy.overrides]]
# module = '*.tests.*'
# ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
